Move Attach/DetachFrame to BrowserTest
